add moved classes
