+++ app/src/main/java/org/guru/playlistmaker/data/search/impl/TrackRepositoryImpl.kt
@@ -6,7 +6,7 @@ import org.guru.playlistmaker.data.search.dto.TrackSearchRequest
 import org.guru.playlistmaker.data.search.dto.TrackSearchResponse
 import org.guru.playlistmaker.domain.search.TrackRepository
 import org.guru.playlistmaker.domain.search.model.Track
-import org.guru.playlistmaker.util.Resource
+import org.guru.playlistmaker.domain.search.Resource
 
 class TrackRepositoryImpl(private val tracksHistoryStorage: TracksHistoryStorage, private val networkClient: NetworkClient) :
     TrackRepository {
