+++ app/src/main/java/org/guru/playlistmaker/domain/SingleLiveEvent.kt
@@ -1,4 +1,4 @@
-package org.guru.playlistmaker.util
+package org.guru.playlistmaker.domain
 
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.MutableLiveData
