+++ app/src/main/java/org/guru/playlistmaker/domain/search/Resource.kt
@@ -1,4 +1,4 @@
-package org.guru.playlistmaker.util
+package org.guru.playlistmaker.domain.search
 
 sealed class Resource<T>(val data: T? = null, val message: String? = null) {
     class Success<T>(data: T): Resource<T>(data)
