+++ app/src/main/java/org/guru/playlistmaker/domain/search/TrackRepository.kt
@@ -1,7 +1,6 @@
 package org.guru.playlistmaker.domain.search
 
 import org.guru.playlistmaker.domain.search.model.Track
-import org.guru.playlistmaker.util.Resource
 
 interface TrackRepository {
     fun searchTracks(expression: String): Resource<List<Track>>
